lib update

+++ lib/lmj/src/containers/container_helpers.hpp
@@ -7,7 +7,7 @@ namespace lmj::detail {
  * @tparam n an unsigned int value
  * @return 0 with the smallest type which can represent n
  */
-template<std::size_t n>
+template<std::uint64_t n>
 consteval auto needed_uint() {
     if constexpr (n <= std::numeric_limits<std::uint8_t>::max()) {
         return std::uint8_t{};
